List of changes: + Added more music commands and moved common cases to set for a faster lookup. + Added a command to toggle special case commands detection. + Added a check for dm to avoid runtime exceptions. +

+++ elfbot.py
@@ -1,7 +1,7 @@
 from discord.ext.commands import Bot, Context, MissingRequiredArgument, MissingPermissions
 from discord import Activity, ActivityType, Message
-from extensions.music_handler import is_music_related, in_music_channel, get_music_channel_id_for_guild_id
-from extensions.prefix_handler import get_prefix, get_prefix_for_guild_id
+from extensions.music_handler import is_music_related, in_music_channel, get_music_channel_id_for_guild
+from extensions.prefix_handler import get_prefix, get_prefix_for_guild
 from utils.utils import *
 
 elfbot = Bot(command_prefix=get_prefix)  # callable prefix - invoked on every message
@@ -37,15 +37,18 @@ async def on_message(message: Message):
     if message.author == elfbot.user:
         return
 
+    if not message.guild:   # DM case
+        return
+
     if elfbot.user.mentioned_in(message):
-        pf = get_prefix_for_guild_id(message.guild.id)
+        pf = get_prefix_for_guild(message.guild.id)
         author = message.author
         await message.channel.send(f'{author.mention}\nMy prefix in this server is {pf}\nUse "{pf}help" for more info')
 
     elif is_music_related(message) and not in_music_channel(message):
         await message.delete()
         log_event(f"<server='{message.guild}'> Caught unauthorized music related message by {message.author}")
-        music_channel = elfbot.get_channel(get_music_channel_id_for_guild_id(message.guild.id))
+        music_channel = elfbot.get_channel(get_music_channel_id_for_guild(message.guild.id))
         if message.embeds:
             for embed in message.embeds:
                 await music_channel.send(embed=embed)
